requirements.txt
keras
tensorflow
numpy
pillow
scikit-image